requirements.txt
python-telegram-bot==22.0
python_dotenv==1.1.0